+++ packages/core/src/configuration.js
@@ -109,7 +109,7 @@ const _read = (configFilePath, projectConfig, flags) => {
 
   target = resolvePath(configPath, target)
   adapters = R.map(resolveAdapters, adapters || {})
-  ui = data.ui || {}
+  ui = data.ui || { base: '/' }
   themes = data.themes || [{ id: '_default', title: 'Default' }]
 
   data = R.assoc('source', source, data)