+++ packages/core/test/builder_test.js
@@ -29,6 +29,9 @@ const state = {
         title: 'Default'
       }
     ],
+    ui: {
+      base: '/'
+    },
     target,
     adapters
   },