+++ packages/core/test/component_test.js
@@ -23,6 +23,9 @@ const state = {
         title: 'Default'
       }
     ],
+    ui: {
+      base: '/'
+    },
     adapters
   }
 }