+++ packages/core/test/variant_test.js
@@ -24,6 +24,9 @@ const state = {
         title: 'Default'
       }
     ],
+    ui: {
+      base: '/'
+    },
     adapters
   }
 }